bridge/requirements.txt
websockets>=12.0
pystray>=0.19.5
Pillow>=10.0.0
